Fixed repeating notification issue
