+++ src/context/index.ts
@@ -1,3 +1,15 @@
 import { AuthContextProvider, useAuthContext } from './AuthContext';
+import { ProductContextProvider, useProductContext } from './ProductContext';
+import { CartContextProvider, useCartContext } from './CartContext';
+import { WishContextProvider, useWishContext } from './WishContext';
 
-export { AuthContextProvider, useAuthContext };
+export {
+  AuthContextProvider,
+  useAuthContext,
+  ProductContextProvider,
+  useProductContext,
+  CartContextProvider,
+  useCartContext,
+  WishContextProvider,
+  useWishContext,
+};
