+++ src/pages/Products.tsx
@@ -1,5 +1,5 @@
 import { ProductCard, Transition, Filters } from '../components';
-import { useProductContext } from '../context/ProductContext';
+import { useProductContext } from '../context';
 
 const Products = () => {
   const { products } = useProductContext();
