+++ src/sass/pages/_products.scss
@@ -14,4 +14,9 @@
     gap: 3rem 2rem;
     padding: 0 1rem;
   }
+
+  &__warning {
+    font-size: $fs-xs;
+    color: $color-red;
+  }
 }
